requirements.txt
ruamel.yaml
pygit2
gitpython